move Project class to session module
