+++ afolder/rapida.json
@@ -1,0 +1,5 @@
+{
+    "name": "afolder",
+    "config_file": "/home/work/py/geo-cb-surge/afolder/rapida.json",
+    "folder": "/home/work/py/geo-cb-surge/afolder"
+}
