+++ pyproject.toml
@@ -45,7 +45,9 @@ dependencies = [
     "matplotlib",
     "folium",
     "jupyterlab_widgets",
-    "ipywidgets"
+    "ipywidgets",
+    "sumpy",
+    "pydantic"
 ]
 
 [project.scripts]
